requirements.txt

fastapi
uvicorn
python-multipart
streamlit
requests
langchain
langgraph
openai
pydantic
qdrant-client
sentence-transformers
python-docx
pdfplumber
pillow
pytesseract
